Api/pyproject.toml
[project]
name = "no-bullshit-timer-api"
version = "0.0.1"
description = "A RESTful api for the timer.reinhardt.ai application"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "attrs==23.2.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "fastapi==0.112.0",
    "greenlet==3.0.3",
    "h11==0.14.0",
    "idna==3.7",
    "iniconfig==2.0.0",
    "packaging==24.1",
    "passlib==1.7.4",
    "pluggy==1.5.0",
    "pydantic==2.8.2",
    "pydantic_core==2.20.1",
    "PyJWT==2.9.0",
    "pytest==8.3.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.31",
    "starlette==0.37.2",
    "typing_extensions==4.12.2",
    "uvicorn==0.30.5",
    "pytest_mock==3.14.0",
    "aiosqlite==0.20.0",
    "asyncpg==0.29.0"
]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
